requirements.txt
flask
flask-socketio
celery
redis
python-dotenv
selenium
webdriver-manager
requests
